Separated the big app.js file into several files
Divided the entire app.ts script into smaller pieces, dividing by functionality. Imported each relevant script into each other and resolved the requireJS issues.
