+++ src/Project.ts
@@ -1,0 +1,18 @@
+//project type
+
+export module Project {
+  export enum ProjectStatus {
+    active,
+    finished,
+  }
+
+  export class Project {
+    constructor(
+      public id: string,
+      public title: string,
+      public description: string,
+      public people: number,
+      public status: ProjectStatus
+    ) {}
+  }
+}
