+++ src/autobind.ts
@@ -1,0 +1,17 @@
+//autobind decorator, this will enable the methods to automatically bind to using 'this' instead of
+//having to do .bind(this) every time
+export function autobind(
+  _target: any,
+  _methodName: string,
+  descriptor: PropertyDescriptor
+) {
+  const originalMethod = descriptor.value;
+  const adjDescriptor: PropertyDescriptor = {
+    configurable: true,
+    get() {
+      const boundFunction = originalMethod.bind(this);
+      return boundFunction;
+    },
+  };
+  return adjDescriptor;
+}
